SPAZA-297 Generic OutOfRange, remove nan treatment
There is no need to mention projections or inputs in `OutOfRange` since
it operates purely on numbers. Simplify the exception class name to
follow suit. I dabbled with the idea to make a single `OutOfRange` class
that doubles as the exception too, but that is maybe a bit too much
overloading and unnecessary parsimony.

The 'nan' treatment turned out to be harder to implement than expected
and more effort than 'clip', so remove it. Users like Mattieu prefer
the 'clip' treatment anyway. The problem is exemplified by the SIN
projection which checks cos(theta) to see whether the target is within
90 degrees of the reference, but then doesn't use cos(theta) in its
calculations of x and y, so making it NaN doesn't help. The hope was
that the NaNs would filter through to the final answer but it doesn't
always. This will require some more workarounds for which I don't
currently have an appetite.

This addresses JIRA tickets SR-1502 and SR-1511.

[v1: leave 'nan' commented out in tests since it will make a comeback]