Revert to sexagesimal in Antenna description
This matches the typical lat-lon-alt input format and is less prone
to numerical fluff and large decimal expansions. The default precision
of the latitude and longitude is limited by Astropy 4.0 to 4 decimal
places at the arcsecond level, which is about 3 mm on the Earth's
surface. Just keep this in mind in case we need higher precision in
future. The height is rounded to the nearest micrometre, so that's OK.

This makes all the tests in test_antenna.py pass again.