Remove troublesome corner case from catalogue test
The Zenith target is constructed with an azimuth of 0 degrees, but
anything pointing straight up has an undefined azimuth... The rising
and setting filter would both pass through an azimuth of 0 as a corner
case, but this broke with Astropy.

First off, the Zenith target had a (valid!) azimuth of 182 degrees.
Then, even if moved to "azel, 0, 89", the Astropy altaz object returned
an azimuth of 359.9999999999862 degrees, instead of 0 or even 360
(which could have been modulo-ed to 0 inside the filter code).

In short, it's a mess. Only add Zenith after the azimuth filter tests
(Zenith can't rise or set anyway!). This reduces the expected number
of targets by 1 in both rising and setting cases.