Fix azel method for stationary bodies
An (az, el) target should return the (az, el) with which it was
constructed when Target.azel() is called. The current version first does
a loop via ICRS and back upon calling the base azel(), which introduces
small numerical errors. Rather specialise `StationaryBody.compute` to
switch the order of azel and radec around.

This fixes failed tests in test_delay.