Finesse the various requirements.txt files
The SARAO Jenkins files (requirements.txt and test-requirements.txt)
need recursive enumeration of all requirements, with pinning of any
packages not in the Docker base requirements. Thankfully pytest is
already in the base.

The system-requirements.txt is presumably a SARAO CAM thing. Update it
for good measure (it does not need recursion).