Turn Timestamp into a shim for Astropy Time
It would have been nice to ditch `Timestamp` completely or at least
make it a subclass of `astropy.time.Time`. There are some issues with
this though, not the least of which is the fact that `Timestamp` does
arithmetic in seconds while `Time` operates in days.

The following changes are made to `Timestamp`:

  - The `Timestamp` object contains an Astropy `Time` object exposed
    via the `time` attribute which does all the work under the hood.
  - The `time` attribute replaces the `to_ephem_date` method as the
    gateway to the underlying functionality.
  - The object can be constructed from a `Time` object or another
    `Timestamp` (a shallow copy).
  - Construction from a string is slightly more stringent, disallowing
    strings that contain a year, a year and month, or a date and hour
    only. These unnatural prefixes are accepted by PyEphem but not
    Astropy.
  - The object can be constructed from an array of floats, strings or
    `Time`s (not covered by unit tests yet).
  - The comparison operators are fleshed out to handle array
    comparisons too (`total_ordering` did not like that...).
  - The `secs` attribute and `to_mjd` method work as before.

Remaining problems are the `local` timezone-aware string method (which
happens to clash with the Astropy namespace too) and a lack of array
testing.

This addresses JIRA ticket SPAZA-153.