Compress projection tests further
Rationalise the corner cases by introducing a sphere->plane->sphere
function which checks the forward and reverse projections at the same
time. Make a parametrized test case for offsets around the origin and
another for offsets around the pole. In the latter case the SIN
projection needs to back off slightly from the maximum offset of 90
degrees to avoid hitting the domain boundary. Setting this backoff to
1e-12 instead of 1e-8 radians also allows the plane->sphere projection
to pass.