Don't strip 0 fractional seconds or crash __str__
When the fractional seconds of a Timestamp is zero, the resulting '.000'
suffix of its string representation was initially stripped to match the
existing katpoint behaviour. With array values this becomes problematic,
since some strings could be stripped while others aren't, leading to
ragged string arrays. The initial solution was to raise an exception
a la __float__, but this is too onerous.

For starters, don't strip the suffix. It's really only the katpoint
tests that complain, and those can be fixed. Anyone else is free to
strip this themselves. This removes the need to test the return type
of `strftime` or loop over it to convert strings.

Rather just turn the array into a string too. While we are at it, use
`iso` instead of `strftime` since this already matches the existing
katpoint format (except for the suffix stripping).