Replicate internals in copy constructor
When initialising a `Timestamp` from another `Timestamp`, instead of
referencing the underlying `Time` object with dubious consequences,
make a replica instead. This shares the immutable internal
representation of `Time` (i.e. the `jd1` and `jd2` members, which make
up the bulk of the object if it's a large array), but copies the
ancillary attributes related to formatting.