Broadcast AltAz coordinates in compute()
In order to compute positions with AltAz, the frame needs to be
complete with obstime and location. Adding these attributes directly
to the internal coord fails, because the scalar (az, el) values won't
broadcast automatically. Therefore first repeat these coordinates to
have the same shape as `obstime` (normally the source of vectorisation)
before replication. Also add a comment to explain why we short-circuit
the transform_to in the case of AltAz->AltAz transformation.