Fix visibility_list display and speed it up
Change the angle display back to sexagesimal.

Since Astropy's (az, el) call tends to be costly (10 ms for a celestial
source and 100 ms for a Solar System object), try to vectorise as much
as possible. The test catalogue has 115 celestial sources and 10
Solar System objects, so each catalogue-wide (az, el) costs 2 seconds...

Similar to geometric_delay, combine the calculation to see if a source
is rising or setting with the actual position calculation, and get rid
of the sort() call, which does a redundant azel(). The number of `azel`
calls drops from 5 to 2 if the fringe rate is also requested, and 1
otherwise, with a corresponding speed-up. The ultimate would be to pool
all the celestial targets into a single SkyCoord...

The test_catalogue tests now takes 26 seconds on my laptop where it
used to take 45 seconds.

There's no need to `import time` in the tests.