More MR fixes
Fix local() fractional seconds. It was broken for values below 0.1,
outputting e.g. 0.12 instead of 0.012. Use the Time output precision
consistently in local(). Add tests for this, also changing the precision
during the test.

Remove workaround for multi-dimensional array of Times/Timestamps.
Rather treat Timestamp and Time as special inputs to Timestamp,
resulting in a straight copy of the underlying time object and avoiding
normalisation via ndarray. It is still possible to crash Astropy with
a multi-dimensional array of objects, but just don't do that :-) Let
Astropy fix the use case instead if it's that important.

Add more string representation tests. Mock better. Describe time format
better in docstring.