Restore target tests that used to fail
Fulfill the dream of having calculated coordinates that *exactly* match
the corresponding ones with which a Target was constructed, for all
relevant body types. So if you have an radec Target, for example,
Target.radec() should return the original coordinates with no need for
a timestamp or antenna. This had issues in the original katpoint due to
limitations in libastro, but was eventually sorted out without updating
the test. Only Galactic still has an issue round-tripping on PyEphem.