Improve error checking of missing antennas
Depending on its body type, some coordinate methods of Target can be
called without an Antenna and others cannot. For example, an "radec"
Target can return radec and galactic without an Antenna but needs one
for azel.

This has to be checked in the underlying Body, which hopefully knows
which frames will require an EarthLocation in compute() and which don't.
Raise a ValueError if the location is missing and it is needed. If you
don't, you typically get a cryptic AttributeError from the bowels of
Astropy.

Add tests to check that the expected methods raise errors (or not) on
various body types.

Some caveats:
  - The apparent radec is currently geocentric but could revert back to
    topocentric at some stage.
  - While a location is mostly needed by AltAz frames, this could
    extend to other bodies in future. The latest Astropy uses the
    location in get_body_barycentric() for light travel times, so it
    might make sense to depend on Antennas in SolarSystemBody too.