Ensure that coordinates match timestamp shape
To maintain compatibility with the original katpoint, it is desirable
to have the SkyCoords output by azel(), radec(), etc have the same shape
as the timestamps. This is generally the case when the output frame
contains an obstime, as with AltAz, GCRS and CIRS. An important
exception is ICRS / Galactic, which has no concept of obstime. In this
case, explicitly repeat the scalar coordinate to match the obstime
shape in compute(). Test that this works for some basic coordinate
methods.

This assumes that the Target coordinate is a scalar (ie we are tracking
a single body) and the Antenna location is a scalar (ie we are not
doing an array at once). These assumptions might have to be revisited
at some stage, but let's start small.