Rework Body into abstract base class
The base compute() had default values for `obstime` and `location` but
the derived `StationaryBody` did not, which breaks the contract.

Rework this so that the base compute() has no default but is also
abstract. The implementation shifts to `FixedBody` (the main derived
class that used it). Make `NullBody` a `FixedBody` to use this too.
The base `Body` now only stores the name, which is better for dynamic
bodies like `SolarSystemBody` and `EarthSatelliteBody` anyway.

Other minor MR fixes:
  - FixedBody is specified to be ICRS, even if it accepts any sky
    coordinate.
  - Don't mention that SolarSystemBody computes GCRS positions (internal
    detail).