Relax checks in anticipation of new Astropy
We will need the latest Astropy for TEME support, which also improves
the Solar System get_body() routine to include the observer in the
calculation of light travel time. This made a 0.3 arcsecond difference
to the position of the Moon.

The coordinate checks are there to verify the katpoint Bodies and not
really to debug Astropy itself, so some tolerance is fine.

Relax the coordinate checks to 1 mas precision in general, and 1 arcsec
for the Moon. This will be tightened again once we depend on Astropy
4.1+.