Check array outputs against corresponding scalars
In order to tighten up the tests, compare the elements in the array
output of Target methods for 2D timestamps against the corresponding
scalar outputs. Introduce a helper function that caters for two
flavours: straight numbers that will be assembled into a grand ndarray
for comparison, and SkyCoords that contain internal arrays and are
better compared via a separation() call (since the basic element is
already a 2D spherical direction). Check the numerical value of
Target.separation because it is easy.

This does add about 7 seconds to the tests on my laptop because of the
overhead of scalar Astropy transforms.