Specify that Body takes Frame/SkyCoord instances
In order to simplify the interface, specify that both Body.coord and
the frame passed to Body.compute can be an instance of either `SkyCoord`
or `BaseCoordinateFrame`. This allows us to assume the presence of
various methods, since those two classes share a large part of their
API. The output of Body.compute is one of these two classes as well,
which makes everything consistent.

The only code change is that we now instantiate ICRS() and Galactic()
instead of passing the classes themselves. This might be very slightly
slower.