Switch to Astropy 4.1rc1 for access to TEME frame
This is the first PyPI release that offers the TEME frame as a builtin
option. It also introduced some changes to the Solar System ephemerides
(topocentric light travel times) and the precision of coordinate tests
had to be adjusted.

Also include a workaround to allow separation() to be called on the
AltAz of a FixedBody vs the AltAz of a SolarSystemBody. The latter
contains additional frame attributes (obsgeoloc and obsgeovel) that
don't contribute to the calculation but crashes an initial frame
equivalency check.