Use Astropy TEME frame instead of Pyorbital
Earth satellites described by the SGP4 model (and distributed as
Two-Line Elements) return positions and velocities relative to the
True Equator, Mean Equinox (TEME) frame, an Earth-centred inertial
frame somewhat like GCRS. This used to be handled by Pyorbital but
since version 4.1 Astropy will have builtin support for this frame
(see astropy/astropy#10149).

Since we don't need the (az, el) of an observer in this calculation
anymore, it is now safe to call `EarthSatelliteBody.compute` without
a location (it defaults to the centre of the Earth). Update the tests
to reflect this.

Add tests against Skyfield which has decent satellite support.
Consolidate the TLE test parameters as module constants in the process.

The new approach is within 0.25 arcseconds of Skyfield, instead of
1.7 arcseconds with the Pyorbital approach.

This addresses JIRA ticket SPAZA-120.