Move readdb inside relevant Bodies
This is a better fit which brings together all EDB and SGP4 related
routines in body.py. Rename readdb/writedb to from_edb/to_edb. Turn
`from_edb` into a class method that serve as alternate constructor for
FixedBody and EarthSatelliteBody.

Rework stars.py to only construct a dict of FixedBodies. There is
nothing left to test. Construct the dict directly in the module,
without the need for a helper function.