Change TLE description to comma-separated lines
The original "tle" body type was an ephemeral format (pun intended)
that was only officially used in unit tests and as intermediate format
in the `Catalogue.add_tle` method. Its description string contains
newlines, which breaks the assumption of one target per line, but
matches the raw format in TLE files ("name \n line1 \n line2 \n").
The archived datasets store TLE targets in XEphem EDB format instead.

Restore this body type to the glory it deserves. Instead of newlines,
store the two lines with a comma in between, which matches azel, radec,
etc and also opens the door for satellite flux models. The old format
is now invalid. The name is found in the usual spot in the string (no
more tle_name...) and if absent, call the satellite "unnamed".

It is an open question whether the TLE description should be one or two
fields. Two fields make it compatible with other body types and have
a natural connection with the two lines of the TLE. However, there is
an XML format looming in the near future of satellite descriptions
(search "Orbit Mean-Elements Message (OMM)") that is a better match for
one field. In the end I've gone with two because it simplifies flux
models. And an XML version could just have an empty second field.

This uses the reasonably fresh sgp4 `export_tle` function which appeared
in version 2.6. However, I missed that the accelerated sgp4init function
needs 2.7 so jump ahead to that.

This addresses JIRA ticket SPAZA-121.