Do TLE epoch check on underlying sgp4 object
Instead of parsing the TLE strings and extracting the epoch and mean
motion ourselves, rely on the parameters stored on the sgp4 Satrec
object. This improves robustness and maintainability.

Expose the TLE epoch on EarthSatelliteBody as an Astropy `Time`, since
we need this in two places by now.

Instantiate the katpoint Targets up front in `add_tle()`, which gives
us access to Satrec via the EarthSatelliteBody. The targets would have
been constructed in `add()` next anyway, with the same level of error
handling.