Rework TLE epoch checks to use Astropy units
Add tests that catch the log messages to ensure that the TLE checks
are done for both near-Earth and deep-space satellites. Add a near-Earth
Iridium satellite TLE to the unit test.

Even fix a bug: the satellite name was stuck at the last one in the
list - should have been `target.name`.