Add tests for angle conversion routines
Before we change these, add tests. I picked test_body.py because
body.py will be their future home (and ephem_extra.py will disappear).

The angle_from_hours routine fails when passed a decimal string. This
should be in degrees and not hours. The existing code circumvents this
by first attempting to convert a decimal string to a float in radians.