Improve angle conversion routines
First attempt to construct an Astropy `Angle` directly and only fall
back to default units if that fails due to missing units. This ensures
that the angle can be a copy of an existing `Angle` object without
coercing it to another unit.

Another change is that decimal strings are always in degrees, while
sexagesimal strings (and tuples) are either in degrees or hours,
depending on the function. Have radians as the fallback unit because
it is stricter (eg it won't accept tuples), and this means that
ndarrays without units will be in radians too, just like scalar numbers.