Consolidate angle conversion routines
The `angle_from_degrees` and `angle_from_hours` functions only differ in
how they handle sexagesimal input, so make that a parameter of a common
`to_angle` function instead. Add a docstring.

Fix up the use of this function in Target. It is not necessary to cast
the angle to a float in radians first. Also promote the Angle to
a Latitude or Longitude where possible.