Move wrap_angle to top level of package
This is a tricky one. The `wrap_angle` function is not used within
katpoint itself, but external scripts use it in all sorts of places.

There is no obvious NumPy equivalent, but in Astropy one would do
`Angle.wrap_at('180d')`. This would require external users to convert
all their float angles to Astropy objects. Removing `wrap_angle` from
the public API would therefore cause some extra pain.

We therefore leave it in for now, but move it out of ephem_extra to
the only sensible place - the top.