Get rid of katpoint.is_iterable
The main use of this helper function is to patch together array-based
and scalar-based code paths inside the main katpoint functions.

Modern NumPy allows most array operations on 0-dimensional arrays, so
a better scheme is to cast scalars to arrays, perform the calculations
along a single array-based code path and turn 0-d ndarrays back into
scalars at the end (if needed) with `ndarray.item()`.

Replace is_iterable with a mix of the 0-d ndarray scheme, attempting
iteration and handling the exception, and Astropy Time array handling.

The refactor of `DelayCorrection.corrections` revealed that the
function crashes if the timestamp is a sequence with a single element,
but this was also the case in the old code. Figuring that out on
another day...

There is very little external use of is_iterable (two places in katdal
come to mind), so it won't be sorely missed.

And poof! ephem_extra.py is gone.

This addresses JIRA ticket SPAZA-124.