Rework Antenna initialisation
Construct an Antenna from an Astropy `EarthLocation` instead of raw
lat-lon-alt. The WGS84 route is still used to combine the delay model's
ENU parameters with the reference location, but this happens internally
and is really meant for description strings. This breaks the "power
user" Antenna initialiser but there are few of those users (and they'll
like the new __init__ more :-)).

This also fixes an egregious bug. The location used the altitude of the
ref_location by mistake, because the one variable was called `elevation`
and the other one was called `height`.

An Antenna is now allowed to have no name (still considering the
ramifications...). This enables the following three main initialisation
routes, aka The API:

  - Antenna('description string')
  - Antenna(another Antenna())
  - Antenna(EarthLocation())

Additionally, split out the description string parser to its own
slightly more efficient factory method:

  - Antenna.from_description('description string')

The __str__ method now returns the description string, which seems
neater than a contrived string that no-one looks at (leave that
one for __repr__). It also removes the need for `format_katcp`,
a MeerKAT-specific anachronism.

Allow additional parameters to override the contents of a description
string or existing Antenna object during initialisation. Previously
this was considered an error, mostly because the first parameter used
to perform double duty as the string `name` or a description string,
and this was an extra check to catch broken names or descriptions.
The new scheme seems more natural.

Turn the position_* tuples into properties to make them read-only.
They all derive from ref_location + delay_model (or location).

The array reference antenna has no need for a diameter or beamwidth.
Use the reference location during construction to preserve accuracy.

Add extra unit tests for Antenna initialisation (to prevent a future
egregious bug).

This addresses JIRA ticket SPAZA-260.