Rework default antenna parameters
It is now possible to override a non-default parameter value with
a default one. For example, you can set the antenna name of an existing
Antenna to ''. Also highlight that an existing Antenna object or
description string effectively provides default values for the antenna
parameters, which can still be overridden during initialisation.