Add skeleton for tropospheric delays
Work within refraction.py instead of starting a new module (simpler).

The tropospheric delay is basically split into a zenith part and a
mapping part (dealing with elevation dependence). Both will depend on
the antenna location. The skeleton deals with parsing a model ID string
into these parts, instantiating the appropriate objects a la Factory
and constructing an internal function to calculate the delay. The class
is read-only, so that the model ID can unambiguously represent the
actual algorithm in use.

The TroposphericDelay class is therefore a glorified function, with some
initial state involving the antenna location that is useful to calculate
just once up front.

This addresses JIRA ticket SPAZA-122.