Require Astropy 4.1
Astropy 4.1 causes some Solar System objects to shift by 5 mas in the
body unit tests. Since the tolerance is 1 mas, this causes the tests
to fail. This is most likely due to astropy/astropy#10666, which
improves the treatment of solar light deflection for nearby objects.
Take the new positions as gospel, since the old behaviour is considered
to be a bug.

Now that Astropy 4.1 is out we can also remove a workaround for 4.1rc1
(fixed in astropy/astropy#10658).