Get rid of DelayCorrection cache
This was meant to save 50% of computations in the live CBF delay update
use case, where we need to calculate the next delay in order to obtain
delay rates and fringe rates, only to compute the same value again on
the next update.

It turns out to be problematic because the cache key is timestamp only,
while the target may change by the next update. This should invalidate
the cache, but doesn't. There could also be a change in offset on the
same target and with the same timestamp, leading to increasingly
complicated cache keys. See USRCH-19 for a discussion of this issue.

These complications are not worth the saving in computation, so get rid
of the infernal cache.