Promote delays() and inputs to public API
The `DelayCorrection` object also serves as a `CorrelatorDelayModel`
which combines all the `DelayModel`s of the underlying antennas in
order to calculate delays efficiently. It is not a `Model` per se,
though, since we don't want to enumerate all the individual parameters
inside. And we want to turn delays into corrections too, which is how
we ended up with `DelayCorrection`.

This publishes the "delay model" part of the class, by making
_calculate_delays and _inputs public instead of private. This allows
users to calculate pure delays without the correction part, which is
more amenable to comparisons with other libraries.