Rework delay timestamp API
Consolidate all timestamps into a single `timestamp` parameter, thereby
getting rid of `next_timestamp`. If you want to pass in more than
one timestamp, put them in an array-valued timestamp instead. A side
effect is that the delay and fringe rates are now always calculated,
with a default time interval of 1 second if a single timestamp is
provided.

In addition, let the `delays` method accept array-valued timestamps
since it will be way more efficient. Update the tests to handle the
API changes on both the input and output of `corrections`.