Redo corrections() output and introduce units
The original idea for the output of the corrections method was that
it will be delay and phase polynomials, initially first-order (offset
and slope). While MeerKAT has never gone beyond first-order, this could
be expanded to cubic and quintic splines. The NCRA team fitted quintic
polynomials to `DelayCorrection` output, while Jive uses Akima cubic
splines with Calc (available in SciPy).

The downside of polynomial coefficients is that they typically have
different units (e.g. s, s / s, rad, rad / s), making it hard to express
as a single Quantity. Here we went the other way, splitting up the
polynomial parts into separate Quantities with well-defined units.

Another option would be to offload the delay and fringe rates (and even
the fringe phase) to external interpolation functions, but for now we
keep it together.

Fix the `delays` method to support arbitrarily-shaped timestamps
(especially scalars and 1-D arrays) and return a Quantity.