Relax restriction on reference antenna
The `DelayCorrection` object expected a list of antennas with the
same reference position, so that the ENU offsets directly serve as
baselines. This forces the array antenna locations to be expressed as
reference + offset, instead of a direct location in geocentric or
geodetic form.

If the reference position of an antenna is the same as that of the
overall `DelayCorrection` object, still use the `DelayModel` as is to
avoid tiny numerical errors creeping in and ending up in description
strings. If not, convert the antenna to ECEF / XYZ and obtain a new ENU
offset relative to the `DelayCorrection` reference position.