Use Astropy units in DelayCorrection
This extends the use of Astropy units to `extra_delay`, `max_delay`
and `sky_centre_freq`, as well as the return type of `delays()`.

Save the centre freq as Hz and the extra delay as seconds in the
JSON description string.

Decompose the unit of the product of `sky_centre_freq` and `delays`
(which should be dimensionless, i.e. turns) to avoid Frankensteins like
"GHz rad s" instead of rad and "GHz rad" instead of rad/s.

`DelayModel` obviously still needs units. I'm postponing this until I
have figured out how to base it on the `astropy.modeling` package.