Depend on Astropy 4.2
The timestamp tests failed because the latest Astropy 4.2 reworked
the Time arithmetic a tiny bit. It was always going to be imprecise
at the 10-20 picosecond level because we are using UTC timestamps and
TAI offsets, and the scale change from UTC to TAI and back introduces
tiny rounding errors (even if the offset is zero, I gather).

This seems like a good time to upgrade to Astropy 4.2, which gives us
`Time.isclose` to check if two times are similar within this error.
We were already anticipating the 4.2 feature of `np.atleast_1d` support
in `DelayCorrection`, so go for it.

Update requirements. And fix the spelling of Astropy in the README.