Fix Astropy 4.2 regression
The shape/size/ndim NumPy functions used to play nicely with
`ShapedLikeNDArray` objects like `Time` and `AltAz` in Astropy 4.1.
This has regressed in Astropy 4.2, which breaks the unit tests. Work
around it for now in the tests by accessing the relevant property
instead if np.shape fails.

See astropy/astropy#8610 for a discussion.