SPAZA-297 Move latitude check into sphere_to_ortho
Most sphere_to_plane routines call `sphere_to_ortho` and then check
the returned cos(theta) for validity. At this point the ortho_x and
ortho_y variables are then passed along as is.

However, if theta is clipped this should also affect (ortho_x, ortho_y)
since the length of this 2-D vector is sin(theta). This will ensure that
the (x, y) plane coordinates are clipped to the boundary of the valid
domain instead of wrapping back inside. If the target point is
diametrically opposite the reference point, it is more tricky to clip
because (x, y) = (0, 0) and we lose the direction from the origin
towards the boundary. In this case we choose an arbitrary point on the
boundary, since this corner case is highly unlikely in practice.

Introduce a `clip_2d` function to change the length of (x, y) safely,
even if the old length is 0. This is a bit more general than required,
as it supports a range of valid radiuses while we typically use it to
force the radius to a new value.

Suppress rounding errors in out-of-range checks. Where relevant, change
the thresholds of checks to be in terms of angles instead of cosine
values. Improve the threshold in one instance (sphere_to_plane_tan)
so that one can now get within 0.2" of the domain boundary. This is an
arbitrary choice that results in (x, y) values of up to ~1e6 (they
blow up at the boundary).

[v1: redid the tests in pytest]