SPAZA-297 Restore 'nan' out-of-range treatment
Now that `safe_scale` is in place, it is almost trivial to re-enable
the replacement of out-of-range values with NaNs. The only function
that needed a bit of help is the rather complex `plane_to_sphere_ssn`,
which has four loosely coupled intermediate values with individual
out-of-range checks. An invalid elevation output therefore did not
guarantee an invalid azimuth output. The solution is to force `az`
to be NaN if `el` is NaN at the end of the function.

Expand the unit tests to check for NaNs in all the expected places,
with additional checks for `plane_to_sphere_ssn`.

[v1: redid the tests in pytest, even adding SSN plane_to_sphere clips]