SPAZA-297 Remove more crutches in OutOfRange.treat
Let the out-of-range check be exact, as that is what one would expect
of the action of `np.clip`. The fuzzy check was introduced to compensate
for rounding errors when calculating np.cos(max_theta) - for example,
because np.cos(np.pi / 2.0) == 6e-17 instead of 0. Rather specify the
thresholds directly in units of cos(theta), which also allows better
correspondence with the original AIPS thresholds.