Rename extra_delay -> extra_correction
This removes the dissonance between the name of the parameter and the
way it is applied. It is subtracted from the output of the delay model
and added to the corrections going to the CBF, so it is fundamentally
a "correction".

It also makes it clear that it is not an intrinsic part of the delay
model per se, but part of the correction machinery instead (just like
the sky centre frequency used to obtain fringe phase).

The DelayCorrection description string parser still accepts the original
name as well, to remain backwards compatible with delay models stored in
datasets.