SPAZA-297 Replace OutOfRange class with functions
The class machinery is a bit convoluted and we also want to introduce
context-local state (threading.local for now, but a context variable
one day). Follow NumPy's approach to floating-point error handling:

  - np.seterr() -> set_out_of_range_treatment()
  - np.geterr() -> get_out_of_range_treatment()
  - np.errstate() -> out_of_range_context()
  - FloatingPointError -> OutOfRangeError
  - treat_out_of_range_values()

These functions and the corresponding OutOfRangeError exception remain
in the projection submodule, since they are considered more advanced
and the extra namespace helps to clarify their use. Don't delete
`projection` like the other submodules to make the functions accessible.

This also resolves the confusion between OutOfRange and OutOfRangeError.

[v1: the projection module is already importable at top level]