Transpose the output of DelayCorrection.delays
Change the output shape from times x inputs to inputs x times. I already
had my reservations about the order in the previous MR, and the upcoming
vectorisation effort for refraction correction seems to work better with
the new shape. It will allow easy broadcasting along timestamps, which
suits the MeerKAT weather measurements recorded per time and not per
antenna.

This is also a better fit to DelayCorrection.corrections, which used to
transpose the delay array back to inputs x times anyway. Both transposes
are now gone.

The delay array is small enough (100 x 1000s of elements mean a few MB)
to fit into memory for efficient conversion to the typical MeerKAT data
shape of time x frequency x baseline, so the new ordering should not
hinder that process.

Call the number of antennas `A` instead of `M` for consistency with
comments in the code (and I remember it better that way!).