New public attribute: locations
This combines the reference location and all A antenna locations into
a single `EarthLocation` of shape (1 + A,), which will simplify and
speed up Astropy calculations by vectorising them. It can be done up
front before calls to `delays()`.

It is a public attribute so that the CALC test can use it without
feeling bad about underscores.