Let Antenna parameters accept EarthLocations
This improves interoperability between katpoint and Astropy. The main
driver is `DelayCorrection`, where it would be beneficial to pass
an `EarthLocation` object with multiple locations to `Target.azel()`
without converting it to an Antenna first.

In the process, refactor the clumsy `Target._normalise_antenna` method
into two bits. The first part (`_astropy_funnel`) converts timestamps
and antennas into their Astropy counterparts while also accepting those
counterparts directly. This is sufficient for most of the basic
coordinate methods. The second part (`_valid_antenna`) returns a
katpoint Antenna like before, but ensures that it is not None and makes
one from an EarthLocation if necessary. This now imports Antenna, which
justifies the split-up of the old `delay` module.