Rewrite DelayCorrection.delays to use locations
This calculates the (az, el) pointing for all antennas (and the
reference antenna) at once, by broadcasting 1-D timestamps and locations
to arrays of shape (n_ants, n_times). The main reason for this is that
we need per-antenna elevations for non-intersecting axis offset (NIAO)
delays and the pending tropospheric delays. Identical elevations with
the same weather and NIAO values would imply that these two delay
components are identical per dish and therefore negligible.

While Astropy 4.2 has much better broadcasting support, we are forced to
extend `time` and `locations` manually to appease Astropy 4.1 while we
still support Python 3.6. Even 4.2 focusses more on obstime broadcasting
than location broadcasting and will need some support. The downside of
this is that scalar times are 3x slower (5 ms -> 14 ms for 85 MeerKAT+
locations), but the use case of many timestamps has similar performance.

The NIAO delay should be more accurate with individual antenna
elevations (up to ~1 mm difference on MeerKAT+'s 20-km baselines).

Add a simple unit test to check that the delays come out with the right
shape for all target body types. This will be useful once we upgrade to
Astropy 4.2 as well.