Fix astrometric (ra, dec)
ICRS is a poor replacement for astrometric (ra, dec) because it is
barycentric instead of geocentric (or even better, topocentric). This
results in wildly different directions for nearby objects: basically
anything in the Solar System and especially Earth satellites.

Rework `Body.compute()` to accept a `to_celestial_sphere` parameter.
If True, find the (az, el) of the body and discard the corresponding
distance, in effect projecting its 3D position vector to infinity onto
the celestial sphere. When this 2D direction vector is transformed to
the final frame, it will remain topocentric.

Also fix the related apparent (ra, dec) and galactic methods. Add notes
that apparent (ra, dec) should really use the TETE frame (only available
in Astropy 4.2). A side effect is that Solar System and Earth satellite
bodies now need Antennas / locations to do astrometric (ra, dec).

Update the unit tests. Compare radec results to PyEphem. Add a funky
unit test that verifies an observed conjunction of Jupiter, Saturn and
the ISS, all done in astrometric (ra, dec). Add a DelayCorrection unit
test that verifies that a zero offset is the same as no offset (the
original spark for this PR!).