SPAZA-267 Add tropospheric delays to delay model
The model description gains a `tropospheric_model` ID string, which
is "None" if no tropospheric corrections will be made. The `delays()`
and `corrections()` methods gain weather parameters (disabled by
default). Since the `delays()` and `corrections()` methods take the same
parameters, let them also share the Parameters section of the docstring.

The current scheme initialises `TroposphericDelay` with the reference
location and then calls it with the elevations of the actual dishes.
This is an approximation with ~0.3 ps errors on MeerKAT+ at 15 degrees
elevation (and ~4 ps at 5 degrees elevation). It also assumes that the
reference location is close to the actual receptors (i.e. not the
centre of the Earth). A more correct approach would be to have multiple
`TroposphericDelay` objects (one per actual receptor location), or to
vectorise `TroposphericDelay` (could be tricky). On the other hand,
these errors might be smaller than the errors introduced by the weather
sensors...