Handle case of FixedBody with 3-D coordinates
Strengthen the gentlemanly agreement and check whether the FixedBody has
2-D or 3-D coordinates. The former case is purely a direction and will
have a norm of 1.0 with a dimensionless unit, while the latter case will
have a distance component in units of length. Check for the unit instead
of doing a float comparison. If 2-D, we are already on the celestial
sphere.

Add a unit test for this. Also add a unit test for StationaryBody which
somehow got left out. Check whether the astrometric (ra, dec) make
sense by converting it back down to (az, el). The non-azel methods
don't raise errors with non-azel targets without antennas anymore,
because it falls back to the geocentre.