Use GCRS for to_celestial_sphere
This has the advantage of proper topocentric support via `obsgeoloc` and
`obsgeovel`, and is particularly simple for Solar System bodies that are
expressed in GCRS anyway. With the latest Astropy master that includes
the astropy/astropy#10994 fix we get astrometric (ra, dec) via GCRS and
AltAz to within 1e-4 mas (or nano-arcseconds) of each other. GCRS-based
astrometric (ra, dec) should be the most accurate version on Astropy 4.1
and 4.2, where it is within 1 mas of Astropy master.

SolarSystemBody to_celestial_sphere is now 2x faster, with azel shifting
by ~0.3" on the Moon (and radec by 40 mas) and much less on the rest.
EarthSatelliteBody is back to the old speed (5x slower), with azel
shifting by 4" (and radec by 0.3"). The 4" is slightly puzzling because
Astropy used to be within 0.5" of Skyfield and is now ~4" out. On the
other hand, it follows the description in astropy/astropy#10994, with
closer sources having progressively worse (az, el).

Adjust the unit test results for Astropy master (4.3.dev331+g960314366).

This addresses SPAZA-289.