Relax the (az, el) checks on Astropy < 4.3
Astropy 4.3 will introduce topocentric CIRS, which improves AltAz
calculations quite a bit. Do all checks on the latest Astropy to 1 mas
accuracy, but relax (az, el) checks on 4.1 and 4.2 according to the
distance of the body from the observer. The (ra, dec) remains within
1 mas for all versions.

This gets worse... Since the delay model is based on (az, el), it is
therefore wrong on nearby objects in Astropy < 4.3. The error is ~20 ps
on the Moon and ~80 ps on a GPS satellite. Ironically, an (ra, dec)
offset target with zero offset ostensibly produces the correct delays,
since astrometric (ra, dec) remains the same for old and new Astropy.