Improve body.compute() API
The main purpose of `body.compute()` is to present a common interface
to `Target` methods that need it. These methods will always have a
timestamp handy (`None` => now) but the Antenna / location may be
missing.

Tweak the `body.compute()` API to reflect this. The `obstime` parameter
is mandatory and we don't attempt to support `obstime=None`. Rather
use `FixedBody.coord` directly to get Galactic coordinates where needed
for description strings (no need to make up a time). The default value
for `to_celestial_sphere` is the same for all base and derived methods.
Shorten the standard calls to body.compute by using normal args instead
of kwargs (it's still very clear).