A script that annotates Great Conjunction photo
The `test_target.py::test_great_conjunction` unit test uses Jason de
Freitas's cool Great Conjunction photo to verify astrometric (ra, dec).
This new script annotates his photo with source positions predicted
from his observation setup. It assumes the Moon and Saturn are in their
expected places and uses that to figure out the rotation and scale of
the photo. It then estimates the positions of Jupiter and the ISS based
on that. It also includes the effects of refraction, given that Jupiter
was only five degrees above the horizon at the time. :-)

This script helps to check for coordinate errors in various versions of
Astropy and katpoint, especially for nearby objects.